Move settings button to uploader
